+++ lib/openstudio/aws/aws.rb
@@ -118,6 +118,7 @@ module OpenStudio
         end
 
         if options[:security_groups].empty?
+          # TODO: Security overhaul: lock down the groups to only specific ports
           # if the user has not specified any security groups, then create one called: 'openstudio-server-sg-v1'
           @os_aws.create_or_retrieve_default_security_group
         else
@@ -139,8 +140,12 @@ module OpenStudio
             user_id: options[:user_id],
             tags: options[:tags],
             subnet_id: options[:subnet_id],
-            associate_public_ip_address: options[:associate_public_ip_address]
+            associate_public_ip_address: options[:associate_public_ip_address],
         }
+
+        # save the worker pem and public to the directory
+        # presently, this will always overwrite the worker key, is that okay? Is this really needed later?
+        @os_aws.save_worker_keys('.')
 
         # if instance_data[:ebs_volume_id]
         #   server_options[:ebs_volume_id] = instance_data[:ebs_volume_id]
@@ -150,8 +155,6 @@ module OpenStudio
 
         @instances_json = instances_json
         save_cluster_json(@instances_json)
-
-
 
         # Print out some debugging commands (probably work on mac/linux only)
         puts ''