+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Version 0.3.0 (Unrealeased)
 * Add method to save the cluster JSON
 * Add top level terminate method to delete the machine that belong to the cluster (based on group uuid)
 * Fix setting the group uuid when loading an existing cluster JSON file
+* Upload public/private key for worker-node communication
 
 Version 0.2.6
 -------------